requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Deprecated==1.2.14
distro==1.9.0
Django==5.0.1
idna==3.7
netaddr==1.2.1
pillow==10.3.0
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.1
wrapt==1.16.0
yookassa==3.1.0
djangorestframework~=3.15.1